requirements.txt
dnspython~=2.7.0
idna~=3.10
httpcore~=1.0.9
httpx~=0.28.1
anyio~=4.9.0
h11~=0.16.0
pip~=24.2
attrs~=25.3.0
protobuf~=6.31.1
grpcio~=1.73.1
grpcio-tools~=1.73.1
typing_extensions~=4.14.1
sniffio~=1.3.1
pytest~=8.4.1
Faker~=37.4.0
Pygments~=2.19.2
execnet~=2.1.1
pytest-xdist~=3.8.0
allure-python-commons~=2.14.3
pluggy~=1.6.0
iniconfig~=2.1.0
certifi~=2025.6.15
pydantic~=2.11.7
email_validator~=2.2.0
pydantic_core~=2.33.2
annotated-types~=0.7.0
packaging~=25.0
setuptools~=80.9.0
referencing~=0.36.2
jsonschema~=4.24.0
jsonschema-specifications~=2025.4.1
allure-pytest~=2.14.3
pydantic-settings~=2.10.1
python-dotenv~=1.1.1
websockets~=15.0.1